examples/server/asgi/django_example/requirements.txt
aioredis==2.0.1
asgiref==3.4.1
async-timeout==4.0.2
bidict==0.21.4
click==8.0.3
Deprecated==1.2.13
Django==4.0.1
gunicorn==20.1.0
h11==0.12.0
httptools==0.3.0
packaging==21.3
pyparsing==3.0.6
python-dotenv==0.19.2
python-engineio==4.3.0
python-socketio==5.5.0
PyYAML==6.0
redis==4.1.0
sqlparse==0.4.2
typing_extensions==4.0.1
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
wrapt==1.13.3
